Moved code back to lib.rs, removed trait bounds on NewConnector trait
